+++ apps/client/src/app/pages/simulator/components/simulation-min-stats-popup/simulation-min-stats-popup.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Simulation } from '@ffxiv-teamcraft/simulator';
 import { TranslateModule } from '@ngx-translate/core';
 import { NzAlertModule } from 'ng-zorro-antd/alert';
+import { NzDividerModule } from 'ng-zorro-antd/divider';
 
 import { DialogComponent } from '../../../../core/dialog.component';
 
@@ -10,13 +11,17 @@ import { DialogComponent } from '../../../../core/dialog.component';
   templateUrl: './simulation-min-stats-popup.component.html',
   styleUrls: ['./simulation-min-stats-popup.component.less'],
   standalone: true,
-  imports: [NzAlertModule, TranslateModule]
+  imports: [NzAlertModule, NzDividerModule, TranslateModule]
 })
 export class SimulationMinStatsPopupComponent extends DialogComponent implements OnInit {
 
   simulation: Simulation;
 
   stats: { control: number, craftsmanship: number, cp: number, found: boolean };
+
+  statsRaw: { control: number, craftsmanship: number, cp: number, found: boolean };
+
+  bonuses: { control: number, cp: number, craftsmanship: number };
 
   thresholds: number[];
 
@@ -26,9 +31,16 @@ export class SimulationMinStatsPopupComponent extends DialogComponent implements
 
   ngOnInit(): void {
     this.patchData();
-    // setTimeout to queue and make sure CD will see it ocne it's done
+    // setTimeout to queue and make sure CD will see it once it's done
     setTimeout(() => {
       this.stats = this.simulation.getMinStats();
+
+      this.statsRaw = {
+        control: this.stats.control - this.bonuses.control,
+        craftsmanship: this.stats.craftsmanship - this.bonuses.craftsmanship,
+        cp: Math.max(180, this.stats.cp - this.bonuses.cp),
+        found: Object.values(this.bonuses).some(s => s > 0),
+      };
     });
   }